Build Tools: "quietmake"
This script wraps GNU Make, providing output pruning
services that reduce the size of the logs, to make it easier
to find and notice interesting things within the logs.

Lines ending with backslash are joined with the next line.

Messages about entering and leaving a directory with nothing
between them are discarded.

Commented-out example lines are included showing how to
add filters to the script to discard selected additional
output lines if you find them to be bulking up the size
of your build logs.

The goal is a build log where any real issues can be clearly
seen reading the log, with just enough other text to allow
monitoring progress during live builds and to identify the
build context for messages in saved log files.