Fix #4: add "fastmake" to build tools.
This does a "quietmake" with parameters added that usually
reduce the wall-clock time for the build.

Directory printing is turned off, as parallel builds with
jobs in different directories (which happens inside this
project when make -j is used) can cause some messages to be
associated with the wrong directory.