Fix #7: Genearlize the Recursive Build
Generalize the parallel recursive build mechanism so that the
implementation is a single block (not a single block for each
recursive target), driven by a list of targets that are recursive
and a list of subdirectories.

Exposes internal targets of the form "dir/target" so that the
build rules in any specific directory can indicate dependencies
between builds; for instance,

    test/cycle: lib/cycle

would indicate that the cycle target should not start building
in the test subdirectory until it finishes in lib. This is mainly
intended to support a dependency between two directories in the
directory list and a target they both implement. Using this to
do other clever things may result in undefined behavior of the
build system (it might work, but it might STOP working).

Also includes some additional build debug support added while
checking that this facility works.