Test::pass() available, consumes supporting text.
I am getting a bit fed up with supporting text from passing
conditions being included as supporting text in subsequent
failing conditions.

Testing code can now call t.pass() to indicate that previously
sent supporting text is associated with this named test condition,
and should not be presented with subsequent conditions.

It is still perfectly fine to run several cases in a row ending
with a failing case, without a t.pass(), if you still want all
of the supporting text for the series to be presented as supporting
text for the failing case ...

In addition, accumulated supporting text will not be dumped into
the body of the XML <test> object when it is closed, as this does
not translate into any useful display.