Bad type decision in C++ "streams" library
For a given iostream s, "s.width()" returns the value
recently given in an "s << setw(n)" directive. Alas,
the type returned by s.width() may have values that
do not fit in the "setw()" parameter.

Somwehere, we need a conversion. Do it when we fetch
the incoming requested field width.