Create new BIST related classes.
Bist::Fun (in BistFun.hh and BistFun.cc)
  corresponds to a test "case" that is a sequence
  of operations, some of which involve evaluating
  whether there has been a failure. This level of
  granularity might correspond to testing one
  method in a class, or one function of a system.

Bist::Pkg (in BistPkg.hh and BistPkg.cc)
  corresponds to a collection of the above tests,
  which might correspond to all the tests for a
  software package.

Granularity can be adjusted to taste; all BIST knows
is that a Pkg has a name, and it has Fun things that
have names, and those have supporting evidence text
and marks for PASS/SKIP/FAIL/ERROR, and the eventual
reporting is going to work on a summary-by-Fun or
perhaps summary-by-Pkg basis.

These classes are set up to run their BIST the first
time an object of their class is created.

BistBist.cc triggers BIST for BistFun and BistPkg, so we
end up with BIST logs for each class plus BistBist itself.

Code coverage is checked with GCOV. This is not a perfect
guarantee that the covered code is tested, but it is fairly
certain that code never visited is not as well tested as
it probably should be.

At least for now, some of the Bist code is a little contorted
so that GCOV can't finger any lines as uncovered, despite the
fact that this test code does not exercise every direction of
every branch -- in the test code that tests the test code.